+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file. For commit 
  - Sidebar links cointains urls with 2x repeated source name (#1847)
  - 2 factor auth getting overwritten if you edit the auth from admin (#1819)
  - New created user's files are listing error, Probably related with language setting. (#1565)
+ - thumbnail generation cpu/memory and concurrency bug.
  
 ## v1.2.1-beta